% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_valids}
\alias{filter_valids}
\title{Filter proteins by valid values}
\usage{
filter_valids(
  df,
  metadata,
  unique_proteins,
  min_prop = NULL,
  at_least_one = FALSE,
  labeltype = 1
)
}
\arguments{
\item{df}{A data frame containing the raw protein expression data.}

\item{metadata}{A data frame mapping intensity column names to groups.}

\item{unique_proteins}{a list containing unique proteins output from obtain_unique_proteins function .}

\item{min_prop}{A numeric threshold for the minimum number of valid
values. For example, 2 means a protein must be valid
in at least 2 samples to be kept.}

\item{at_least_one}{Considers whether the condition must be met with both or only one condition.}

\item{labeltype}{Type of mass spectrometry experiment.}
}
\value{
A list containing the filtered data frame ("filtered") and
the LOG2-transformed data frame ("LOG2").
}
\description{
A helper function to filter a protein expression data frame
(e.g., from MaxQuant) based on a minimum number of valid
(non-NA, non-zero) values across samples.
}
\examples{

raw_df <- data.frame(
  Protein = c("P1", "P2", "P3", "P4"),
  LOG2_A1 = c(20, 21, NA, 22),
  LOG2_A2 = c(20, NA, NA, 22),
  LOG2_B1 = c(25, 24, 23, NA),
  LOG2_B2 = c(25, 24, 23, 26)
)
meta_df <- data.frame(
  intensity_sample_name = c("A1", "A2", "B1", "B2"),
  group = c("Control", "Control", "Treatment", "Treatment"),
  log2_col = c("LOG2_A1", "LOG2_A2", "LOG2_B1", "LOG2_B2")
)
unique_list <- list(
  data.frame(Protein = character()),
  data.frame(Protein = character())
)

filtered_data <- filter_valids(
  df = raw_df,
  metadata = meta_df,
  unique_proteins = unique_list,
  min_prop = 0.5,
  at_least_one = FALSE,
  labeltype = 1
)

head(filtered_data)

}
