% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gostplot_func}
\alias{gostplot_func}
\title{Manhattan Plot for Functional Enrichment}
\usage{
gostplot_func(terms, ...)
}
\arguments{
\item{terms}{A list of length 3 generated by Goterms_finder function.}

\item{...}{Additional arguments passed to \code{gprofiler2::gostplot}}
}
\value{
A ggplot or plotly object representing the Manhattan
plot of functional enrichment.
}
\description{
This function is a helper wrapper that extracts the raw
gprofiler2 results (the third element) from the output of
the Goterms_finder function to generate a Manhattan plot of enriched terms.
}
\examples{
\donttest{
gostplot_func(res)
}
}
