% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{histogram}
\alias{histogram}
\title{Plot a Histogram}
\usage{
histogram(df, colname, color, title)
}
\arguments{
\item{df}{A data frame.}

\item{colname}{The name of the (numeric) column to plot.}

\item{color}{The color for the histogram.}

\item{title}{The main title for the plot.}
}
\value{
A histogram plot.
}
\description{
A simple wrapper to plot a histogram for a specific column
in a data frame.
}
\examples{
my_data <- data.frame(p.value = rnorm(100))
histogram(my_data, "p.value", "lightblue", "P-Value Distribution")


}
