% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{identify_proteins}
\alias{identify_proteins}
\title{Identify Proteins}
\usage{
identify_proteins(raw, metadata, platform, selected_conditions)
}
\arguments{
\item{raw}{A data frame containing the raw protein expression data.}

\item{metadata}{A data frame mapping intensity column names to groups.}

\item{platform}{An indicator (e.g., 1) for the data platform (e.g., MaxQuant).}

\item{selected_conditions}{Conditions compared from you experiment.}
}
\value{
A list containing three data frames: \code{unique_df1} (proteins unique
to df1), \code{unique_df2} (proteins unique to df2), and \code{common_df} (proteins
common to both).
}
\description{
Identifies unique and common proteins between two data frames.
}
\examples{
  metadata_example <- data.frame(
     intensity_sample_name = c("Log_C1", "Log_C2", "Log_T1", "Log_T2"),
     group = c("Control", "Control", "Treatment", "Treatment")
  )

  raw_example <- data.frame(
     Protein = c("Prot_A", "Prot_B", "Prot_C", "Prot_D"),
     Log_C1 = c(25.1, 24.2, 0, 22.0),
     Log_C2 = c(25.0, 24.5, 0, 22.2),
     Log_T1 = c(0, 26.1, 23.5, 21.9),
     Log_T2 = c(0, 26.0, 23.8, 22.1)
  )
  plot_result <- identify_proteins(
      raw = raw_example,
      metadata = metadata_example,
      platform = 1,
      selected_conditions = c("Control", "Treatment")
  )
  print(plot_result)


}
