% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{impute_KNN_data}
\alias{impute_KNN_data}
\title{Impute missing data using k-NN}
\usage{
impute_KNN_data(df, LOG2.names, ...)
}
\arguments{
\item{df}{A numeric matrix with NA values.}

\item{LOG2.names}{A character vector of column names (from \code{df}) to impute.}

\item{...}{Additional arguments passed to the \code{plot} function.}
}
\value{
A numeric matrix with NA values imputed.
}
\description{
Uses K-Nearest Neighbors (k-NN) to impute missing (NA)
values in a log-expression matrix.
}
\examples{
\donttest{
log_matrix_na <- matrix(
    c(
        10.0, 10.2, NA, 10.5, 10.1, 9.9, # Sample 1 (with NAs)
        12.0, NA, 11.9, 12.1, 11.8, NA
    ), # Sample 2 (with NAs)
    nrow = 6, ncol = 2,
    dimnames = list(
        c("P1", "P2", "P3", "P4", "P5", "P6"),
        c("Sample1", "Sample2")
    )
)

LOG2.names <- c("Sample1", "Sample2")

imputed_matrix <- impute_KNN_data(as.data.frame(log_matrix_na), LOG2.names)

print(imputed_matrix)
}
}
