% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interactions_down}
\alias{interactions_down}
\title{Plot STRING Interactions for Down-regulated Proteins}
\usage{
interactions_down(df, taxonid, score)
}
\arguments{
\item{df}{A data frame from statistical analysis (e.g., limma), must have
'expression' and 'Protein' columns.}

\item{taxonid}{Numeric. The NCBI taxonomy ID for the species (e.g., 9606
for human).}

\item{score}{Numeric. The minimum required interaction score (0-1000).}
}
\value{
This function is called for its side effect of generating a plot.
It does not return an object.
}
\description{
Maps down-regulated proteins to the STRING database and
plots the network.
}
\examples{
\donttest{
if (requireNamespace("STRINGdb")) {
  # Create synthetic data
  limma_df <- data.frame(
    Protein = c("P10415", "P62258"), # CALCA, ACTB
    expression = c("Down-regulated", "Down-regulated")
  )
  # Run function
  interactions_down(limma_df, taxonid = 9606, score = 400)
}
}

}
