% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obtain_LOG.names}
\alias{obtain_LOG.names}
\title{Obtain LOG2 Column Names}
\usage{
obtain_LOG.names(df)
}
\arguments{
\item{df}{A data.frame to search.}
}
\value{
A character vector of column names.
}
\description{
A helper function to find all column names containing "LOG2".
}
\examples{
# Create a small example data frame
my_df <- data.frame(
  Protein = c("P1", "P2"),
  Intensity.A = c(100, 200),
  LOG2.A = c(6.6, 7.6),
  Intensity.B = c(110, 210),
  LOG2.B = c(6.7, 7.7)
)

# This example is runnable!
log_names <- obtain_LOG.names(my_df)
print(log_names)

}
