% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{postimputation_state}
\alias{postimputation_state}
\title{Plot Post-Imputation Density}
\usage{
postimputation_state(data_filtered, imputation, LOG2.names)
}
\arguments{
\item{data_filtered}{The data frame \emph{before} imputation, containing
non-finite values (NA, Inf).}

\item{imputation}{An indicator (1 or 2) specifying the imputation method.
1 = \code{impute_KNN_data}, 2 = \code{impute_data}.}

\item{LOG2.names}{A character vector of the LOG2 column names to use.}
}
\value{
A ggplot object showing the two density plots.
}
\description{
Plots the density of intensity values before and after
imputation to visualize the effect of the imputation.
}
\examples{
# This function depends on `impute_KNN_data` and `impute_data`
# We must create those functions and the data they need.

# 1. Create synthetic data
data_filt <- data.frame(
  Protein = c("A", "B", "C", "D"),
  LOG2.S1 = c(10, 12, NA, 8),
  LOG2.S2 = c(11, 13, 9, NA),
  KEEP = c(TRUE, TRUE, TRUE, TRUE) # For impute_data
)

# 2. Create minimal versions of the imputation functions for the example
impute_KNN_data <- function(df, lnames, k) {
   df[lnames][is.na(df[lnames])] <- mean(unlist(df[lnames]), na.rm = TRUE)
   return(df)
}
impute_data <- function(df, lnames) {
   df[lnames][is.na(df[lnames])] <- rnorm(sum(is.na(df[lnames])), mean = 7, sd = 0.3)
   return(df)
}

# 3. Run the plot function
postimputation_state(data_filt, imputation = 2, LOG2.names = c("LOG2.S1", "LOG2.S2"))

}
