% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preimputation_state}
\alias{preimputation_state}
\title{Plot Missingness Pattern}
\usage{
preimputation_state(df, cond.names)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{cond.names}{A character vector of column names from \code{df} to be
included in the missingness plot.}
}
\value{
A plot showing the aggregation of missing values.
}
\description{
Generates a missing value aggregation plot from the \code{VIM}
package to visualize patterns of missing data.
}
\examples{
# Create synthetic data with NAs
log_df <- data.frame(
  LOG2.C1 = c(10, 10.1, NA, 9.9),
  LOG2.C2 = c(11, NA, NA, 11.2),
  LOG2.T1 = c(12, 12.1, 12.2, 12.3)
)

# Plot the missingness pattern
preimputation_state(log_df, c("LOG2.C1", "LOG2.C2", "LOG2.T1"))


}
