% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qqplot_function}
\alias{qqplot_function}
\title{Create a Q-Q Plot}
\usage{
qqplot_function(df, colname1, colname2, color)
}
\arguments{
\item{df}{A data frame.}

\item{colname1}{Character. The name of the column for the x-axis.}

\item{colname2}{Character. The name of the column for the y-axis.}

\item{color}{The color for the points.}
}
\value{
A ggplot object.
}
\description{
Generates a Q-Q plot to compare the distributions of two
columns in a data frame.
}
\examples{
my_data <- data.frame(
  Theoretical = rnorm(100),
  Sample = rnorm(100, mean = 0.5)
)
qqplot_function(my_data, "Theoretical", "Sample", "blue")

}
