% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_traianprot.R
\name{run_traianprot}
\alias{run_traianprot}
\title{Ejecuta el pipeline completo de TraianProt}
\usage{
run_traianprot(
  file_path_raw,
  file_path_metadata,
  output_directory,
  selected_conditions,
  file_path_report = NULL,
  platform_id = 1,
  organism_id = 1,
  labeltype = 1,
  norm_method = "mean",
  imputation_method = "3",
  min_prop_filtering = 0.5,
  min_unique_peptides = 1,
  test = 2,
  psms = TRUE,
  paired = FALSE,
  statval = 1,
  way = 2,
  logfcup = 1,
  logfcdown = -1,
  sig = 0.05,
  conditions = NULL,
  id_target = "ENSG",
  organismo = NULL,
  threshold = 0.05,
  custombg = FALSE,
  label = 3,
  font_size = 12,
  terms_number = 12,
  taxonid = 9606,
  score = 400
)
}
\description{
Ejecuta el pipeline completo de TraianProt
}
