% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatterplot_function}
\alias{scatterplot_function}
\title{Create a Scatterplot with Marginal Densities}
\usage{
scatterplot_function(df, colname1, colname2)
}
\arguments{
\item{df}{A data frame.}

\item{colname1}{Character. The name of the column for the x-axis.}

\item{colname2}{Character. The name of the column for the y-axis.}
}
\value{
A \code{ggExtra} plot object.
}
\description{
Uses \code{ggplot2} and \code{ggExtra} to create a scatterplot comparing
two columns, with marginal density plots.
}
\examples{
my_data <- data.frame(
  Sample1 = rnorm(100),
  Sample2 = rnorm(100)
)
scatterplot_function(my_data, "Sample1", "Sample2")

}
