% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsne}
\alias{tsne}
\title{t-SNE Plot}
\usage{
tsne(x, metadata, perplexity_num, selected_conditions)
}
\arguments{
\item{x}{A data frame or matrix with expression data. Columns are samples
(e.g., LOG2 intensity values). Rows are features (e.g. proteins).}

\item{metadata}{A data frame with sample metadata. Must have 'group' and
'log2_col' columns.}

\item{perplexity_num}{Numeric. The perplexity value for the t-SNE algorithm.}

\item{selected_conditions}{A character vector of length 2 specifying the two
groups from the 'group' column to compare.}
}
\value{
A ggplot object.
}
\description{
Generates a t-SNE plot for sample clustering using ggplot2.
}
\examples{
  df <- data.frame(
     Sample1 = c(10, 11, 12, 10),
     Sample2 = c(10, 11, 12, 11),
     Sample3 = c(20, 21, 22, 20),
     Sample4 = c(20, 21, 22, 21)
  )
  row.names(df) <- c("P1", "P2", "P3", "P4")

  # 2. Metadata
  metadata <- data.frame(
     intensity_sample_name = c("Sample1", "Sample2", "Sample3", "Sample4"),
     group = c("A", "A", "B", "B")
  )

  # IMPORTANT: For small datasets, perplexity must be low (usually < samples/3)
  # We use perplexity = 1 for this tiny example.
  tsne_plot(df, metadata, perplexity = 1)

}
