% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{venn_diagram}
\alias{venn_diagram}
\title{Create a Venn Diagram}
\usage{
venn_diagram(df, unique_proteins, color1, color2)
}
\arguments{
\item{df}{A data frame of common proteins (must have a "Protein" column).}

\item{unique_proteins}{A list containing two data frames:
1. Proteins unique to the first set.
2. Proteins unique to the second set.}

\item{color1}{Color for the first set.}

\item{color2}{Color for the second set.}
}
\value{
A gList object (a plot) representing the Venn diagram.
}
\description{
Generates a Venn diagram plot object from a list of unique
proteins and a data frame of common proteins.
}
\examples{
common_proteins <- data.frame(
    Protein = c("ProteinA", "ProteinB", "ProteinC")
)

unique_protein_list <- list(
    Control = data.frame(Protein = c("ProteinD", "ProteinE")),
    Treatment = data.frame(Protein = c("ProteinF", "ProteinG", "ProteinH", "ProteinI"))
)

my_venn_plot <- venn_diagram(
    df = common_proteins,
    unique_proteins = unique_protein_list,
    color1 = "blue",
    color2 = "red"
)

grid::grid.newpage()
grid::grid.draw(my_venn_plot)

}
