% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dotplot_func}
\alias{dotplot_func}
\title{Create Dot Plot from Enrichment Results}
\usage{
dotplot_func(terms, ...)
}
\arguments{
\item{terms}{A list, where the first element (\code{terms[[1]]}) is a
\code{compareClusterResult} object.}

\item{...}{Additional arguments passed to \code{enrichplot::dotplot}.}
}
\value{
A ggplot object.
}
\description{
Generates a dot plot from a \code{compareClusterResult} object,
faceted by condition.
}
\examples{
# This example requires a valid compareClusterResult object.
# We will create a minimal synthetic one.
if (requireNamespace("clusterProfiler") && requireNamespace("methods")) {
  dummy_data <- data.frame(
    Cluster = "Up-regulated", Category = "BP", ID = "GO:0006915",
    Description = "apoptosis", p.value = 0.01, adj.P.Val = 0.01,
    query_size = 50, Count = 5, term_size = 100,
    effective_domain_size = 10000, geneID = "P04637/P00533",
    GeneRatio = "5/50", BgRatio = "100/10000", Conditions = "Up-regulated"
  )
  dummy_cc_result <- methods::new("compareClusterResult",
                                compareClusterResult = dummy_data)
  terms_list <- list(dummy_cc_result)
  dotplot_func(terms_list)
}

}
