% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{identify_proteins}
\alias{identify_proteins}
\title{Identify Proteins}
\usage{
identify_proteins(raw, metadata, platform, selected_conditions)
}
\arguments{
\item{df1}{A data frame for condition 1, must have a "Protein" column.}

\item{df2}{A data frame for condition 2, must have a "Protein" column.}
}
\value{
A list containing three data frames: \code{unique_df1} (proteins unique
to df1), \code{unique_df2} (proteins unique to df2), and \code{common_df} (proteins
common to both).
}
\description{
Identifies unique and common proteins between two data frames.
}
\examples{
df1 <- data.frame(Protein = c("P1", "P2", "P3"), Value = 1:3)
df2 <- data.frame(Protein = c("P2", "P3", "P4"), Value = 2:4)
id_list <- identify_proteins(df1, df2)
print(id_list$common_df)
print(id_list$unique_df1)

}
