% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quick_filtering}
\alias{quick_filtering}
\title{Quick Data Filtering}
\usage{
quick_filtering(
  raw,
  platform,
  organism,
  metadata,
  selected_conditions,
  directory_path = NULL
)
}
\arguments{
\item{raw}{A data frame of raw proteomics data (e.g., from MaxQuant).}

\item{platform}{An indicator (e.g., 1) for the data platform (e.g., MaxQuant).}

\item{organism}{An indicator (e.g., 1 for Human, 2 for Mouse, etc.).}

\item{metadata}{A data frame mapping intensity column names to sample names.}

\item{selected_conditions}{(Not currently used in function) A vector of
selected conditions.}

\item{directory_path}{(Not currently used in function) A path to a directory.}
}
\value{
A data frame filtered to remove contaminants/reverse hits, with
new "Protein" and "Protein_description" columns.
}
\description{
Filters raw proteomics data (e.g., from MaxQuant) to remove
contaminants, reverse hits, and site-only identifications. It also parses
protein identifiers from Fasta headers.
}
\examples{
 raw_data <- data.frame(
  Potential.contaminant = c("", "+", "", ""),
  Reverse = c("", "", "+", ""),
  Only.identified.by.site = c("", "", "", "+"),

  Fasta.headers = c(
    ">sp|P04637|P53_HUMAN Cellular tumor antigen p53 OS=Homo sapiens",
    ">sp|P00000|CONTAM_Protein Common contaminant",
    ">sp|P12345|REV_Protein Reverse Sequence",
    ">sp|P62258|1433E_HUMAN 14-3-3 protein epsilon"
  ),
  # Intensity columns
  Intensity.S1 = c(10000, 500, 200, 100),
  Intensity.S2 = c(12000, 600, 150, 90),
  stringsAsFactors = FALSE
)

metadata <- data.frame(
  intensity_sample_name = c("Intensity.S1", "Intensity.S2"),
  sample_name = c("Sample_A", "Sample_B")
)

filtered_df <- quick_filtering(
   raw = raw_data,
   platform = 1,
   organism = 2,
   metadata = metadata,
   selected_conditions = NULL,
   directory_path = NULL
)

}
