% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTraianProt.R
\name{runTraianProt}
\alias{runTraianProt}
\title{Launch the TraianProt Shiny Application}
\usage{
runTraianProt()
}
\value{
A \code{shiny.appobj} object representing the TraianProt app.
}
\description{
This function builds and returns the Shiny app object for TraianProt.
}
\examples{
app <- runTraianProt()

if (interactive()) {
    shiny::runApp(app)
}
}
