% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runTraianProt}
\alias{runTraianProt}
\title{Launch the TraianProt Shiny Application}
\usage{
runTraianProt()
}
\value{
Opens the TraianProt Shiny app in the default web browser.
}
\description{
This function launches the TraianProt interactive Shiny application.
It provides a user-friendly graphical interface to run the proteomics
downstreaming analysis workflow.
}
\examples{
\donttest{

runTraianProt()
}
}
