% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runTraianProt}
\alias{runTraianProt}
\title{Launch the TraianProt Shiny Application}
\usage{
runTraianProt()
}
\value{
Nothing. This function is called for its side effect of
launching the Shiny application.
}
\description{
Launches the Shiny application for the TraianProt package.
}
\details{
This function finds the Shiny app directory within the installed
package and runs it.
}
\examples{
\dontrun{
  # This is the correct way to show an app example
  runTraianProt()
}
}
