% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traian_to_SE}
\alias{traian_to_SE}
\title{Export TraianProt results to SummarizedExperiment}
\usage{
traian_to_SE(df.F, total_dataset, metadata, df_limma)
}
\value{
A \code{SummarizedExperiment} object containing the raw and normalized intensity matrices (assays), sample information (colData), and differential expression statistics (rowData).
}
\description{
Compiles the raw, normalized, and statistical results from
the TraianProt pipeline into a Bioconductor-standard SummarizedExperiment object.
}
\examples{
df_raw <- data.frame(
    Protein = paste0("Protein_", 1:5),
    WT_1 = runif(5, 100, 1000), WT_2 = runif(5, 100, 1000),
    Treat_1 = runif(5, 100, 1000), Treat_2 = runif(5, 100, 1000)
)

df_norm <- df_raw
df_norm[, 2:5] <- log2(df_norm[, 2:5]) # Simulamos la normalización

meta <- data.frame(
    Condition = c("WT", "WT", "Treatment", "Treatment"),
    log2_col = c("WT_1", "WT_2", "Treat_1", "Treat_2")
)

stats <- data.frame(
    Protein = paste0("Protein_", 1:5),
    logFC = c(1.2, -0.5, 2.1, 0.1, -1.8),
    p_value = c(0.01, 0.4, 0.005, 0.8, 0.02)
)

se_object <- traian_to_SE(df_raw, df_norm, meta, stats)
se_object

}
