% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traianprot_power_curve}
\alias{traianprot_power_curve}
\title{Create power curve}
\usage{
traianprot_power_curve(
  df,
  log2_cols,
  foldchange,
  replicatespower,
  alpha_level_choice,
  alpha_level
)
}
\arguments{
\item{df}{A data.frame or matrix containing the expression data.}

\item{log2_cols}{Character vector or numeric indices indicating which
columns of \code{df} contain the abundance values (must be numeric).}

\item{foldchange}{Numeric. The expected Fold Change to calculate the power
(e.g., 2 for doubling, 0.5 for halving).}

\item{replicatespower}{Numeric. Number of replicates per group.}

\item{alpha_level_choice}{Numeric. Choice to consider either p-value or adjusted p-value.}

\item{alpha_level}{Numeric. Global significance level before
correction (defaults to 0.05).}
}
\value{
A \code{ggplot} object containing the power curve.
}
\description{
Calculates the theoretical power curve based on the observed variability
in the experimental data and a desired Fold Change.
}
\examples{
# Simulated data
data <- data.frame(
    ID = letters[1:10],
    R1 = rnorm(10), R2 = rnorm(10), R3 = rnorm(10)
)
plot <- traianprot_power_curve(data, c("R1", "R2", "R3"), foldchange = 2, replicatespower = 5, alpha_level_choice = 1, alpha_level = 0.05)
print(plot)
}
